pyproject.toml
[project]
name = "event-analytics-hub-ingest"
version = "0.1.0"
description = "The Event Analytics Hub Ingest Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-migrate>=4.0.7",
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
]
